+++ config/prezet.php
@@ -54,7 +54,7 @@ return [
 
         'config' => [
             'heading_permalink' => [
-                'html_class' => 'mr-2 scroll-mt-12',
+                'html_class' => 'prezet-heading',
                 'id_prefix' => 'content',
                 'apply_id_to_heading' => false,
                 'heading_class' => '',
@@ -84,9 +84,9 @@ return [
     |
     | Configure how image tags are handled when converting from markdown.
     |
-    | 'path' specifies the route for serving images.
     | 'widths' defines the various widths for responsive images.
     | 'sizes' indicates the sizes attribute for responsive images.
+    | 'zoomable' determines if images are zoomable.
     */
 
     'image' => [
@@ -96,5 +96,21 @@ return [
         ],
 
         'sizes' => '92vw, (max-width: 1024px) 92vw, 768px',
+
+        'zoomable' => true,
     ],
+
+    /*
+    |--------------------------------------------------------------------------
+    | Sitemap
+    |--------------------------------------------------------------------------
+    | The sitemap origin is used to generate absolute URLs for the sitemap.
+    | An origin consists of a scheme/host/port combination, but no path.
+    | (e.g., https://example.com:8000) https://www.rfc-editor.org/rfc/rfc6454
+    */
+
+    'sitemap' => [
+        'origin' => env('PREZET_SITEMAP_ORIGIN', env('APP_URL')),
+    ],
+
 ];