+++ routes/prezet.php
@@ -5,7 +5,7 @@ use BenBjurstrom\Prezet\Http\Controllers\IndexController;
 use BenBjurstrom\Prezet\Http\Controllers\OgimageController;
 use BenBjurstrom\Prezet\Http\Controllers\SearchController;
 use BenBjurstrom\Prezet\Http\Controllers\ShowController;
-use Illuminate\Foundation\Http\Middleware\ValidateCsrfToken;
+use Illuminate\Foundation\Http\Middleware\VerifyCsrfToken;
 use Illuminate\Session\Middleware\StartSession;
 use Illuminate\Support\Facades\Route;
 use Illuminate\View\Middleware\ShareErrorsFromSession;
@@ -13,9 +13,11 @@ use Illuminate\View\Middleware\ShareErrorsFromSession;
 Route::withoutMiddleware([
     ShareErrorsFromSession::class,
     StartSession::class,
-    ValidateCsrfToken::class,
+    VerifyCsrfToken::class,
 ])
     ->group(function () {
+        Route::get('prezet/search', SearchController::class)->name('prezet.search');
+
         Route::get('prezet/img/{path}', ImageController::class)
             ->name('prezet.image')
             ->where('path', '.*');
@@ -30,6 +32,4 @@ Route::withoutMiddleware([
         Route::get('prezet/{slug}', ShowController::class)
             ->name('prezet.show')
             ->where('slug', '.*'); // https://laravel.com/docs/11.x/routing#parameters-encoded-forward-slashes
-
-        Route::get('prezet/search', SearchController::class)->name('prezet.search');
     });