+++ src/Commands/InstallCommand.php
@@ -4,6 +4,7 @@ namespace BenBjurstrom\Prezet\Commands;
 
 use Illuminate\Console\Command;
 use Illuminate\Filesystem\Filesystem;
+use Illuminate\Support\Facades\Process;
 
 class InstallCommand extends Command
 {
@@ -36,6 +37,8 @@ class InstallCommand extends Command
             $this->copyTailwindFiles();
             $this->installNodeDependencies();
 
+            // run in separate process so config changes above are applied
+            Process::run('php artisan prezet:index --force');
             $this->info('Prezet has been successfully installed!');
 
             return self::SUCCESS;
@@ -82,7 +85,7 @@ class InstallCommand extends Command
     protected function copyContentStubs(): void
     {
         $sourceDir = __DIR__.'/../../stubs/prezet';
-        $destinationDir = storage_path('prezet');
+        $destinationDir = base_path('prezet');
 
         if (! $this->files->isDirectory($sourceDir)) {
             $this->warn('Skipping content stubs: source directory already exists.');
@@ -103,7 +106,7 @@ class InstallCommand extends Command
     protected function installNodeDependencies(): void
     {
         $this->info('Installing node dependencies');
-        $packages = 'alpinejs @tailwindcss/forms @tailwindcss/typography autoprefixer postcss tailwindcss';
+        $packages = 'alpinejs @tailwindcss/forms @tailwindcss/typography autoprefixer postcss tailwindcss vite-plugin-watch-and-run';
 
         if (file_exists(base_path('pnpm-lock.yaml'))) {
             $bin = 'pnpm';
@@ -127,8 +130,6 @@ class InstallCommand extends Command
             return;
         }
         $this->info('Adding prezet database');
-        $this->files->copy(__DIR__.'/../../stubs/prezet.sqlite', base_path('prezet.sqlite'));
-
         $configFile = config_path('database.php');
         $config = file_get_contents($configFile);
         if (! $config) {
@@ -137,7 +138,7 @@ class InstallCommand extends Command
             return;
         }
 
-        $diskConfig = "\n        'prezet' => [\n            'driver' => 'sqlite',\n            'database' => base_path('prezet.sqlite'),\n            'prefix' => '',\n            'foreign_key_constraints' => true,\n        ],";
+        $diskConfig = "\n        'prezet' => [\n            'driver' => 'sqlite',\n            'database' => base_path('prezet/prezet.sqlite'),\n            'prefix' => '',\n            'foreign_key_constraints' => true,\n        ],";
 
         $disksPosition = strpos($config, "'connections' => [");
         if ($disksPosition !== false) {
@@ -164,7 +165,7 @@ class InstallCommand extends Command
             return;
         }
 
-        $diskConfig = "\n        'prezet' => [\n            'driver' => 'local',\n            'root' => storage_path('prezet'),\n            'throw' => false,\n        ],";
+        $diskConfig = "\n        'prezet' => [\n            'driver' => 'local',\n            'root' => base_path('prezet'),\n            'throw' => false,\n        ],";
 
         $disksPosition = strpos($config, "'disks' => [");
         if ($disksPosition !== false) {