+++ src/Http/Controllers/SearchController.php
@@ -2,13 +2,12 @@
 
 namespace BenBjurstrom\Prezet\Http\Controllers;
 
-use BenBjurstrom\Prezet\Data\FrontmatterData;
 use BenBjurstrom\Prezet\Data\HeadingData;
 use BenBjurstrom\Prezet\Models\Heading;
+use Exception;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
-use Illuminate\View\View;
 
 class SearchController
 {
@@ -23,9 +22,13 @@ class SearchController
         }
 
         $query = $request->input('q');
+        if (! is_string($query)) {
+            throw new Exception('Query must be a string');
+        }
 
         $results = Heading::where('text', 'LIKE', "%{$query}%")
             ->with('document')
+            ->whereRelation('document', 'draft', false)
             ->limit(5)
             ->get()
             ->map(function ($heading) {