+++ stubs/vite.config.js
@@ -1,5 +1,7 @@
-import { defineConfig } from 'vite';
+import path from 'path'
 import laravel from 'laravel-vite-plugin';
+import { defineConfig } from 'vite';
+import { watchAndRun } from 'vite-plugin-watch-and-run'
 
 export default defineConfig({
     plugins: [
@@ -11,5 +13,14 @@ export default defineConfig({
             ],
             refresh: true,
         }),
+        watchAndRun([
+            {
+                name: 'prezet:index',
+                watch: path.resolve('prezet/**/*.(md|jpg|png|webp)'),
+                run: 'php artisan prezet:index',
+                delay: 1000,
+                // watchKind: ['add', 'change', 'unlink'], // (default)
+            }
+        ])
     ],
 });