+++ tests/TestCase.php
@@ -4,6 +4,8 @@ namespace BenBjurstrom\Prezet\Tests;
 
 use ArchTech\SEO\SEOServiceProvider;
 use BenBjurstrom\Prezet\PrezetServiceProvider;
+use Illuminate\Support\Facades\Artisan;
+use Illuminate\Support\Facades\Route;
 use Orchestra\Testbench\TestCase as Orchestra;
 
 class TestCase extends Orchestra
@@ -31,13 +33,27 @@ class TestCase extends Orchestra
             'throw' => true,
         ]);
 
+        $dbPath = __DIR__.'/../stubs/prezet.sqlite';
+        touch($dbPath);
         config()->set('database.connections.prezet', [
             'driver' => 'sqlite',
-            'database' => __DIR__.'/../stubs/prezet.sqlite',
+            'database' => $dbPath,
             'prefix' => '',
             'foreign_key_constraints' => true,
         ]);
 
         config()->set('database.default', 'prezet');
+
+        Route::group([], function () {
+            require __DIR__.'/../routes/prezet.php';
+        });
+
+        $migrations = __DIR__.'/../database/migrations';
+        Artisan::call('migrate:fresh', [
+            '--path' => $migrations,
+            '--database' => 'prezet',
+            '--realpath' => true,
+            '--no-interaction' => true,
+        ]);
     }
 }