+++ CHANGELOG.md
@@ -2,6 +2,117 @@
 
 All notable changes to this project will be documented in this file.
 
+## [v0.21.0](https://github.com/benbjurstrom/prezet/compare/v0.20.1...v0.21.0) - 2025-01-23
+
+### What's Changed
+
+* 116 add force flag to prezet index command by @benbjurstrom in https://github.com/benbjurstrom/prezet/pull/118
+* 120 remove fresh from index creation by @benbjurstrom in https://github.com/benbjurstrom/prezet/pull/121
+* 122 add force flag to migration by @benbjurstrom in https://github.com/benbjurstrom/prezet/pull/123
+
+**Full Changelog**: https://github.com/benbjurstrom/prezet/compare/v0.20.1...v0.21.0
+
+## [v0.20.1](https://github.com/benbjurstrom/prezet/compare/v0.20.0...v0.20.1) - 2024-12-23
+
+### What's Changed
+
+* 109 update zoomable by @benbjurstrom in https://github.com/benbjurstrom/prezet/pull/110
+
+**Full Changelog**: https://github.com/benbjurstrom/prezet/compare/v0.20.0...v0.20.1
+
+## [v0.20.0](https://github.com/benbjurstrom/prezet/compare/v0.19.0...v0.20.0) - 2024-12-22
+
+### What's Changed
+
+* 106 watch for file changes with vite by @benbjurstrom in https://github.com/benbjurstrom/prezet/pull/107
+* 105 use custom sitemap generator by @benbjurstrom in https://github.com/benbjurstrom/prezet/pull/108
+
+**Full Changelog**: https://github.com/benbjurstrom/prezet/compare/v0.19.0...v0.20.0
+
+## [v0.19.0](https://github.com/benbjurstrom/prezet/compare/v0.18.0...v0.19.0) - 2024-12-21
+
+### What's Changed
+
+* 102 add zoomable images by @benbjurstrom in https://github.com/benbjurstrom/prezet/pull/103
+* 101 ignore drafts in search by @benbjurstrom in https://github.com/benbjurstrom/prezet/pull/104
+* Fixes #97 - issues with special characters such as ' in headings by @djped in https://github.com/benbjurstrom/prezet/pull/98
+
+**Full Changelog**: https://github.com/benbjurstrom/prezet/compare/v0.18.0...v0.19.0
+
+## [v0.18.0](https://github.com/benbjurstrom/prezet/compare/v0.17.0...v0.18.0) - 2024-11-04
+
+### What's Changed
+
+* #88 remove hardcoded route by @djped in https://github.com/benbjurstrom/prezet/pull/89
+* fix: allow images to start with http/https by @timkley in https://github.com/benbjurstrom/prezet/pull/92
+* chore: fix filename of OgImageController by @timkley in https://github.com/benbjurstrom/prezet/pull/90
+* add-x-cloak by @benbjurstrom in https://github.com/benbjurstrom/prezet/pull/94
+
+### New Contributors
+
+* @djped made their first contribution in https://github.com/benbjurstrom/prezet/pull/89
+* @timkley made their first contribution in https://github.com/benbjurstrom/prezet/pull/92
+
+**Full Changelog**: https://github.com/benbjurstrom/prezet/compare/v0.17.0...v0.18.0
+
+## [v0.17.0](https://github.com/benbjurstrom/prezet/compare/v0.16.0...v0.17.0) - 2024-10-16
+
+### What's Changed
+
+* 85 add loading indicator to search by @benbjurstrom in https://github.com/benbjurstrom/prezet/pull/86
+
+**Full Changelog**: https://github.com/benbjurstrom/prezet/compare/v0.16.0...v0.17.0
+
+## [v0.16.0](https://github.com/benbjurstrom/prezet/compare/v0.15.0...v0.16.0) - 2024-10-16
+
+### What's Changed
+
+* 61 add search by @benbjurstrom in https://github.com/benbjurstrom/prezet/pull/83
+* 63 move prezet folder to root by @benbjurstrom in https://github.com/benbjurstrom/prezet/pull/84
+
+**Full Changelog**: https://github.com/benbjurstrom/prezet/compare/v0.15.0...v0.16.0
+
+## [v0.15.0](https://github.com/benbjurstrom/prezet/compare/v0.14.0...v0.15.0) - 2024-09-27
+
+### What's Changed
+
+* 🐛 Fix #76: file extension handling by @fuzzyfox in https://github.com/benbjurstrom/prezet/pull/77
+* 78 fix database malformed error by @benbjurstrom in https://github.com/benbjurstrom/prezet/pull/79
+
+### New Contributors
+
+* @fuzzyfox made their first contribution in https://github.com/benbjurstrom/prezet/pull/77
+
+**Full Changelog**: https://github.com/benbjurstrom/prezet/compare/v0.14.0...v0.15.0
+
+## [v0.14.0](https://github.com/benbjurstrom/prezet/compare/v0.13.0...v0.14.0) - 2024-09-19
+
+### What's Changed
+
+* 71 remove unnecessary type check  by @benbjurstrom in https://github.com/benbjurstrom/prezet/pull/75
+
+**Full Changelog**: https://github.com/benbjurstrom/prezet/compare/v0.13.0...v0.14.0
+
+## [v0.13.0](https://github.com/benbjurstrom/prezet/compare/v0.12.0...v0.13.0) - 2024-09-19
+
+### What's Changed
+
+* 71 Laravel 10 compatibility by @benbjurstrom in https://github.com/benbjurstrom/prezet/pull/72
+* 70 add optional filename parameter to getnav by @benbjurstrom in https://github.com/benbjurstrom/prezet/pull/73
+* 71 Use config front matter by @benbjurstrom in https://github.com/benbjurstrom/prezet/pull/74
+
+**Full Changelog**: https://github.com/benbjurstrom/prezet/compare/v0.12.0...v0.13.0
+
+## [v0.12.0](https://github.com/benbjurstrom/prezet/compare/v0.11.0...v0.12.0) - 2024-09-08
+
+### What's Changed
+
+* 64 run migrations on index by @benbjurstrom in https://github.com/benbjurstrom/prezet/pull/65
+* 66 utf 8 encoding issue by @benbjurstrom in https://github.com/benbjurstrom/prezet/pull/68
+* 64 run migrations on index by @benbjurstrom in https://github.com/benbjurstrom/prezet/pull/69
+
+**Full Changelog**: https://github.com/benbjurstrom/prezet/compare/v0.11.0...v0.12.0
+
 ## [v0.11.0](https://github.com/benbjurstrom/prezet/compare/v0.10.0...v0.11.0) - 2024-09-05
 
 ### What's Changed